product

+++ src/main/java/se/vestige_be/dto/request/CategoryRequest.java
@@ -18,5 +18,5 @@ public class CategoryRequest {
     @Size(max = 500, message = "Description cannot exceed 500 characters")
     private String description;
 
-    private Integer parentCategoryId;
+    private Long parentCategoryId;
 }
