+++ src/main/java/se/vestige_be/repository/CategoryRepository.java
@@ -6,12 +6,12 @@ import se.vestige_be.pojo.Category;
 import java.util.List;
 import java.util.Optional;
 
-public interface CategoryRepository extends JpaRepository<Category,Integer> {
+public interface CategoryRepository extends JpaRepository<Category,Long> {
     Optional<Category> findByName(String name);
 
     boolean existsByName(String name);
 
-    List<Category> findByParentCategory_CategoryId(Integer parentCategoryId);
+    List<Category> findByParentCategory_CategoryId(Long parentCategoryId);
 
     List<Category> findByParentCategoryIsNull();
